requirements.txt
openai~=0.27.0
python-telegram-bot[all]~=20.1